explode formatter hash to case statement
Building the full formatter hash with all parameters
is wasteful, as typical invocation won't include
the majority of the paratemeters in their formats.

So we check which parameters actually occur in the
passed header and format templates and use a
case in a loop over these ones to insert their values.

$ seq 5000000 | range.rb .. -n
- before: 18 s
- after: 11s

$ seq 5000000 | range.rb .. -f '%{IDX}:%{idx} %{file}:%<lno>2d %{match}: %{line}' --header '%{NL}# %{fno}. %{path}'
- before: 26 s
- after: 20s