fix leak upon sequential regexp matches
The insertion of ad hoc ranges representing
regexp match neighborhoods to range list
did not respect order. This can break the
logic of the cleanup routine of the list
which sweeps out old entries with the
assumption of the list being ordered.

However, this was not just simply overlooked;
the idea was that we can be loose here, as the
cleanup routine can deal with some fuzziness:
if an initial segment of the range list is
below the current threshold, it will be swept
out, regardless of the order of the items in it.
So as we progress, we move beyond the given
ad hoc entry and it will be eventually swept
out, its ordering renitence is salvaged.

This unenvenness:
                                   o
                                  oo
                                 oo
                       ooo      o
                     ooo oo   ooo
                    oo    ooooo
                   oo
                  oo
                 oo
+-------------->XX

is overcame when we move beyond the
local peak:

                                   o
                                  oo
+------------------------------->XX
                       XXX      X
                     XXX XX   XXX
                    XX    XXXXX
                   XX

Hwoever, if adjoint match ranges are inserted in
sequence, the new ad hoc items will act like a dam
and protects all the preceding ad hoc entries:

+---------------------> ooo
                          ooo
                            oo
                              oo
                               oo

later...

+--------------------->o
                        oo
                         oo
                          oo
                           ooo
                             oo
                              ooo
                                ooo
                                  oo
                                    oo
                                     oo

So we get a leak.

This could be demonstrated with

$ seq 500000 | range.rb -g . > /dev/null

It wasn't triggered with

$ seq 500000 | range.rb -g '[02468]$' > /dev/null

where matches are not adjoint.